requirements.txt
boto3==1.26.59
Django==4.1.5
django-storages==1.13.2
gunicorn==20.1.0
django-environ==0.10.0
psycopg2-binary==2.9.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.21.4
whitenoise==5.3.0

coreapi==2.3.3
djangorestframework-simplejwt==5.2.1
s3transfer==0.6.1

